Separated utilities and scoring functions into their own files
