+++ variable_importance.py
@@ -14,12 +14,14 @@ correct type from the ctypes module.
 
 @author G. Eli Jergensen <gelijergensen@ou.edu>
 """
-import numpy as np
+from collections import deque
+from ctypes import c_double, c_char_p
 from multiprocessing import Pool, cpu_count
 from multiprocessing.sharedctypes import RawArray
-from ctypes import c_double, c_char_p
-from collections import deque
+import numpy as np
 import sys
+
+import scorers
 
 # These are global variables which are inherited by compute_score_for_column
 inputs_share = None
@@ -70,7 +72,7 @@ def permutation_selection_variable_importances(model, classes, testing_input, te
         subsamples = int(len(testing_output) * subsamples)
     # Check if the score_fn is None
     if score_fn is None:
-        score_fn = accuracy_scorer
+        score_fn = scorers.accuracy_scorer
     # Determine how many variables to return at the end
     if nimportant_variables is None:
         nimportant_variables = len(testing_input[0])
@@ -275,200 +277,6 @@ def compute_score_for_column(c, important_columns, model, classes, npermute, sub
         scores_for_col.append(score)
 
     return np.average(scores_for_col, axis=0)
-
-
-"""Each of these functions with 'scorer' at the end can be used for the score_fn parameter of 
-permutation_selection_variable_importances. Or you can use your own"""
-
-
-def accuracy_scorer(new_predictions, truths, classes):
-    """Scores the model based on the accuracy of the new predictions
-
-    :param new_predictions: The predictions of the model after the permutation
-    :param truths: The true labels of these data
-    :param classes: NOT USED
-    :returns: the accuracy of the new_predictions
-    """
-
-    new_accuracy = float([new_predictions[idx] == truths[idx]
-                          for idx in range(len(truths))].count(True)) / float(len(truths))
-
-    return new_accuracy
-
-
-def gerrity_skill_scorer(new_predictions, truths, classes):
-    """Determines the gerrity skill score, returning a scalar
-
-    :param new_predictions: The predictions of the model after the permutation
-    :param truths: The true labels of these data
-    :param classes: an ordered set for the label possibilities
-    :returns: a single value for the gerrity skill score
-    """
-    # get the confusion matrices
-    new_cm = confusion_matrix(new_predictions, truths, classes)
-    new_table = MulticlassContingencyTable(
-        table=new_cm, n_classes=len(classes), class_names=classes)
-
-    return new_table.gerrity_score()
-
-
-def peirce_skill_scorer(new_predictions, truths, classes):
-    """Determines the peirce skill score, returning a scalar
-
-    :param new_predictions: The predictions of the model after the permutation
-    :param truths: The true labels of these data
-    :param classes: an ordered set for the label possibilities
-    :returns: a single value for the peirce skill score
-    """
-    # get the confusion matrices
-    new_cm = confusion_matrix(new_predictions, truths, classes)
-    new_table = MulticlassContingencyTable(
-        table=new_cm, n_classes=len(classes), class_names=classes)
-
-    return new_table.peirce_skill_score()
-
-
-def heidke_skill_scorer(new_predictions, truths, classes):
-    """Determines the heidke skill score, returning a scalar
-
-    :param new_predictions: The predictions of the model after the permutation
-    :param truths: The true labels of these data
-    :param classes: an ordered set for the label possibilities
-    :returns: a single value for the heidke skill score
-    """
-    # get the confusion matrices
-    new_cm = confusion_matrix(new_predictions, truths, classes)
-    new_table = MulticlassContingencyTable(
-        table=new_cm, n_classes=len(classes), class_names=classes)
-
-    return new_table.heidke_skill_score()
-
-
-def categorical_metascorer(scoring_fn, category=None, selection_strategy=None):
-    """This is a meta-scorer which converts a binary-class scorer to a multi-class scorer for use above using a 
-    one-versus-rest strategy or another specified strategy
-
-    :param scoring_fn: function which is a binary-class scorer (like bias). Must be of the form:
-        (new_predictions, truths, classes, particular class) -> float
-    :param category: the identity of the class to consider (if doing one-versus-rest)
-    :param selection_strategy: either "maximimum", "minimum", "average", or a callable
-        NOTE: if neither category or selection_strategy is specified, prints a warning and defaults to average
-        callable must be of the form (list of scores) -> float
-        category is ignored if selection_strategy is specified
-    :returns: scoring function which wraps correctly around scoring_fn
-    """
-    # First determine whether we are doing ovr or a specified strategy
-
-    if category is None and selection_strategy is None:
-        print("WARNING: categorical_metascorer defaulting to averaging")
-        selection_strategy = 'average'
-
-    if 'max' in selection_strategy:
-        selection_strategy = np.max
-    elif 'min' in selection_strategy:
-        selection_strategy = np.min
-    elif 'avg' in selection_strategy or 'average' in selection_strategy:
-        selection_strategy = np.average
-    else:
-        assert callable(
-            selection_strategy), "ERROR: strategy must be 'minimize', 'maximize', 'average' or a callable"
-
-    def cat_scorer(new_predictions, truths, classes):
-        if selection_strategy is None:  # then we know category isn't None
-            return scoring_fn(new_predictions, truths, classes, category)
-        else:
-            all_scores = [scoring_fn(
-                new_predictions, truths, classes, bin_class) for bin_class in classes]
-            return selection_strategy(all_scores)
-
-    return cat_scorer
-
-
-"""This is a utility function for helping score things. You may also find it useful when 
-implementing your own scoring functions"""
-
-
-def confusion_matrix(predictions, truths, classes):
-    """Computes the confusion matrix of the predictions vs truths
-
-    :param predictions: model predictions
-    :param truths: true labels
-    :param classes: an ordered set for the label possibilities
-    :returns: a numpy array for the confusion matrix
-    """
-
-    matrix = np.zeros((len(classes), len(classes)), dtype=np.float32)
-    for i, c1 in enumerate(classes):
-        for j, c2 in enumerate(classes):
-            matrix[i, j] = [p == c1 and t == c2 for p,
-                            t in zip(predictions, truths)].count(True)
-    return matrix
-
-
-"""This class is borrowed wholesale from the hagelslag repository as of 11 June 2018. It is reproduced here
-with permission. Party responsible for copying: G. Eli Jergensen
-
-@author David John Gagne <djgagne@ou.edu>"""
-
-
-class MulticlassContingencyTable(object):
-    """
-    This class is a container for a contingency table containing more than 2 classes.
-    The contingency table is stored in table as a numpy array with the rows corresponding to forecast categories,
-    and the columns corresponding to observation categories.
-    """
-
-    def __init__(self, table=None, n_classes=2, class_names=("1", "0")):
-        self.table = table
-        self.n_classes = n_classes
-        self.class_names = class_names
-        if table is None:
-            self.table = np.zeros((self.n_classes, self.n_classes), dtype=int)
-
-    def __add__(self, other):
-        assert self.n_classes == other.n_classes, "Number of classes does not match"
-        return MulticlassContingencyTable(self.table + other.table,
-                                          n_classes=self.n_classes,
-                                          class_names=self.class_names)
-
-    def peirce_skill_score(self):
-        """
-        Multiclass Peirce Skill Score (also Hanssen and Kuipers score, True Skill Score)
-        """
-        n = float(self.table.sum())
-        nf = self.table.sum(axis=1)
-        no = self.table.sum(axis=0)
-        correct = float(self.table.trace())
-        return (correct / n - (nf * no).sum() / n ** 2) / (1 - (no * no).sum() / n ** 2)
-
-    def gerrity_score(self):
-        """
-        Gerrity Score, which weights each cell in the contingency table by its observed relative frequency.
-        :return:
-        """
-        k = self.table.shape[0]
-        n = float(self.table.sum())
-        p_o = self.table.sum(axis=0) / n
-        p_sum = np.cumsum(p_o)[:-1]
-        a = (1.0 - p_sum) / p_sum
-        s = np.zeros(self.table.shape, dtype=float)
-        for (i, j) in np.ndindex(*s.shape):
-            if i == j:
-                s[i, j] = 1.0 / (k - 1.0) * \
-                    (np.sum(1.0 / a[0:j]) + np.sum(a[j:k-1]))
-            elif i < j:
-                s[i, j] = 1.0 / (k - 1.0) * (np.sum(1.0 / a[0:i]
-                                                    ) - (j - i) + np.sum(a[j:k-1]))
-            else:
-                s[i, j] = s[j, i]
-        return np.sum(self.table / float(self.table.sum()) * s)
-
-    def heidke_skill_score(self):
-        n = float(self.table.sum())
-        nf = self.table.sum(axis=1)
-        no = self.table.sum(axis=0)
-        correct = float(self.table.trace())
-        return (correct / n - (nf * no).sum() / n ** 2) / (1 - (nf * no).sum() / n ** 2)
 
 
 if __name__ == "__main__":
